broken world and movement lotsa of fixing in progress

+++ RPG.Combat.Kata.Code/Characters/MeleeCharacter.cs
@@ -3,7 +3,7 @@ namespace RPG.Combat.Kata
     public class MeleeCharacter : Character
     {
         
-        public MeleeCharacter(int health = CharacterConstants.MaxHealth, int level = CharacterConstants.DefaultStartingLevel, double speed = 8) : base(health, level, speed)
+        public MeleeCharacter(int health = CharacterConstants.MaxHealth, int level = CharacterConstants.DefaultStartingLevel, int speed = 8) : base(health, level, speed)
         {   
             AttackRange = 2;     
         }
